Extract CPIO as a reusable module
